Re-adjust Arctic Armor Peroperties
Let's try this...  instead of a flat 5 points per armor, let's boost the ambient temperature by 10% when not in the sun, and by 25% when in the sun.  Should make arctic biomes during the day not so bad, but the nighttime still a bit chilly.  The total boost should be 140% ambient temperature in the shade/at night, and 200% in the sun.