Add some blocks that can be grabbed by tornados
Anything not heavy and tied down to the ground should be able to be lifted.  Stairs and slabs are not included (for now) due to it being unnecessary.  If there's a desire to add them as well, I will.

Besides, between this and zombies breaking basic blocks, players will have a hard enough time surviving in their homes.