Renaming back to ItemArmor.cfg
Enviromine keeps wanting to generate a file with this name, which includes defaults for vanilla armor.  It's not a problem in dev since I can just discard changes, but what happens when someone else downloads the pack, starts it up, and gets the incorrectly generated config file?  I'll see what renaming it to the intended file does, and hopefully it won't cause much grief...