Removed Iron Items from Stronghold Corridor Chests
While yes, the player should have iron by now, there is a (very low) chance they happen upon an exposed stronghold early on and gain access to iron too early.  Also, free iron armor? eh, not really, at least not now.  Also the pickaxe is worthless anyways and will most likely just be melted down for iron, and the sword could be a bit of a reward, but smarter players - especially those who haven't gotten iron yet - will see it as an opportunity to melt it down for iron for their pick.