Moved Saltwater Mobs to TF
If there even are any saltwater biomes in the TF...  considering saltwater is bad to drink, adding mobs to the oceans only further enforces the need to not go there.  If mobs spawn in freshwater, but not saltwater, the player has a choice to make; drink saltwater without much opposition, or drink fresh water at the risk of being attacked.