Re-adjust the food formula
Instead of just dropping 20% everytime after the first 5, let's make this a bit more gradual, while also not being so harsh on those first few days of hunting/gathering for food, while also making sure that the higher foods drop off quicker, and aren't necessarily such an obviously better source of food.

This looks very complicated, but here's the concept:

The more full you currently are (hunger and/or saturation) the less a piece of food will restore.  The more food in your history, the more worthwhile a piece of food that gives higher base food will be, whereas morsels won't give much of anything.  The thought process here is that if you're starving, you're more than likely to receive the full benefits from any piece of food, whereas if you're full, you won't get as much.  Over time, however, there'll be diminishing returns in general, although food shouldn't entirely be drained of it's usefulness.  This should make simple foods worth it early on, and more filling foods worth it later on, with an emphasis on eating more when you're hungry, and less when you're full.

NOTE: THIS WILL PROBABLY HAVE TO BE TWEAKED.