Food Formula Attempt #2!
Not as crazy as the last one, but still pretty crazy.

Basically, the hungrier you currently are versus how much the item (or food group the item belongs to) you're trying to eat determines the base amount.  That gets adjusted by the number of times that food or food group has been consumed within the food history, and the end result is something like the previous formula, just less steep of a drop-off.

I'd personally like to have the player's current hunger state be the driving force behind how much a piece of food gives, then determine how much the player has eaten of a particular food or group, so that they can't survive on berries forever, and eating while full does nothing.

I want a dynamic where the player tries to craft better, more complex and filling foods, but has to eat them when they're actually hungry, and not just topping off their hunger bar.