Disable all TiC Tools, Re-Balance Melting Progression
This is a huge change, and will probably come with a fair share of bugs, but it's good for a first draft.  This basically disables any and all TiC tools from being created, in favor of their "vanilla" counterparts.  This should allow me to add Enviromine properties to them, as well as keep a tighter grip on the pack's balance, since tools won't be so easily exploited.

This is all part of #71 and #82.