Remove TheBetweenlands
It's just too large of a mod to keep in the pack, stretching upwards of 80MBs in the version I have, and 200+ in the version that is current at the time of writing this.  I just simply can't keep up with it, and it was underutilized to begin with, so sadly it just has to go.  Maybe in another pack.