Tweak SoL Configurations
Set a max eating duration (2x the default so that food doesn't take obscenely long to consume).

Tweaked food formula, again.  It takes into account most of the same things as before, but now it also takes into account the food groups the food is a part of, as well as how varied the player's diet is.

Lowered food history length down to 64 from 256, as the new formula isn't reliant on it and it just got a bit silly to be so long.