Switch Healing from Food Paradigms
Way back when I adopted some BnB tweaks, I inadvertently made it so that you never regenerated by being full, but instead only healed when you ate food, which is fine in BnB and early versions of TWBB, but now that farming and obtaining food is more of a hassle, and that eating food is more about keeping up with a varied diet rather than cycling through foods to keep up the 100%, this doesn't make any sense anymore.  Granted, regeneration will be 50% that of vanilla, but you'll start to regenerate health at half full, and you'll still regenerate slow the less health you have.

This should close #131.