Update FSP
The best part about the update is the new modular drill heads, but having an updated version is always good.

One thing to note, is that now the steam tools are only harvest level of 0 by default, and the new drill head will allow the Steam Drill to be upgraded to higher harvest levels.  with TWBB Tweaks' new version, the tools will only cost aluminum, instead of steel, so any drill will be an upgrade, rather than a downgrade or being useless.