Add BetterBeginnings Tweaks Script
First up, rebalancing seared and scorched bricks, which will close #139.

This will make it so that Seared (Nether) Brick will require 8 normal clay/nether bricks and a bucket of lava.  If using a normal iron bucket to craft the seared brick, it will be returned, unlike the clay buckets.

Scorched Brick will now require Grout, which was the original way to obtain Seared Brick.  This means that technically it's easier to obtain the High Oven, which will allow more metals to be melted down and processed, but alloying won't happen until Seared Bricks can be made.