Various Recipe Tweaks
Tweaked some book recipes, removed blank wooden patterns (Closes #144), and fixed an erroneous recipe where 2 bread could be made from 3 wheat in a brick oven; it's now 1 for 1 dough.  The SoL food journal now requires a book & quill (along with the achievement book, btw) to keep with the theme of writing in the book.