Tweak Spice of Life Formula (#154)
* Tweak SoL Configurations

Set a max eating duration (2x the default so that food doesn't take obscenely long to consume).

Tweaked food formula, again.  It takes into account most of the same things as before, but now it also takes into account the food groups the food is a part of, as well as how varied the player's diet is.

Lowered food history length down to 64 from 256, as the new formula isn't reliant on it and it just got a bit silly to be so long.

* Change Rounding Mod to Ceiling

That way food will always give at least a half a haunch of hunger.  I know TWBB is supposed to be difficult, but eating something that gives absolutely no benefit isn't hard, it's annoying.