Add Starting Inventory
I need to give the Achievement book, but I'm also thinking of starting players off with 16 coagulated blood to both help ease up on those first few days (it won't matter much after that because of SoL anyways), and because it's TWBB, why not give the player some blood?

This closes #155.